coreboot: re-configure gm45 thinkpads from scratch
TSEG Stage Cache enabled again, because disabling it
did not affect S3 in any way.

Many configs have changed, and debug level is set to 7.

In testing with V-T60 on IRC, it wasn't just removal of
the DDR2 patch that I did, but I re-did the configs too,
in exactly the same way I've done them here, when testing
on an X200 to fix boot issues.

Libreboot does not use defconfigs, instead it uses full
configs, and these have to be updated. I normally just
run make-oldconfig on every config, for revision updates.

However, every now and then, we need to re-do them.

Play it safe and re-do every config. I've double- and
triple-checked that the configs are correct.

Signed-off-by: Leah Rowe <leah@libreboot.org>