grub: avoid printing empty error messages
this replaces the previous behaviour, which erred
on a specific value of grub_errno, which was a
problem if other types of errors used that value.

due to the way i patch out the prefix error messages,
this new patch ensures that only those errors are
silenced. all other messages will be printed.

Signed-off-by: Leah Rowe <leah@libreboot.org>