coreboot: update hp elitebook configs
i had to run make-oldconfig on all of them, because
of the port that riku added the other day. lbmk doesn't
use defconfigs, it uses full configs, so we have to
make sure they're kept in sync

this patch is the result of running the following command
in a fresh clone of lbmk:

./update trees -u coreboot

i should probably switch to defconfigs.

Signed-off-by: Leah Rowe <leah@libreboot.org>