update coreboot configs
this was done automatically by running:

./update trees -u coreboot

this has to be done when adding patches for now board ports,
because of the way lbmk and also coreboot's build systems work.

the configs just have to be re-generated to include a line
that says the entry for the newly added boards isn't set. look
at the diff of this commit as an example.

Signed-off-by: Leah Rowe <leah@libreboot.org>