coreboot/haswell: make IOMMU a runtime option
It is now possible to disable the IOMMU on Haswell
boards, by doing this on your ROM image:

./nvramtool -C libreboot.rom -w iommu=Disable

To enable it again, do this:

./nvramtool -C libreboot.rom -w iommu=Enable

If not specified, the default behaviour is *on*.
A follow-up patch will turn IOMMU *off* by default,
on Dell OptiPlex 9020 SFF/MT, by setting it as such
in cmos.default. This is to make graphics cards work
properly to work around a bug when it's turned on.

Leaving the IOMMU enabled is recommended, if it works.
It works in most cases, including on 9020 SFF/MT when
using the Intel GPU without a graphics card inserted.

Signed-off-by: Leah Rowe <leah@libreboot.org>