coreboot/dell9020*_12mb: Disable IOMMU by default
Needed to make graphics cards work. Turn it on if you're using
only the Intel GPU.

With IOMMU *enabled*, graphics cards do not work reliably at all.
The cause still needs to be investigated, but the symptoms are
graphical corruption on the screen, and Xorg usually crashes.
In some cases (on some cards), TTYs can still be used; the payload
can still be used reliably, on a graphics card, but Xorg fails to
work properly.

This could be a bug in Linux drivers, instead of anything that
coreboot does (not yet tested in factory BIOS).

Leaving it off by default will ensure reliable operation on all
setups, whether an iGPU or dGPU is used.

Signed-off-by: Leah Rowe <leah@libreboot.org>